test: update score calculator with double instead of int scoring for better precision

+++ src/main/kotlin/org/example/Main.kt
@@ -76,7 +76,7 @@ fun main() {
     }
 }
 
-private fun printScores(scores: Map<org.example.domain.Node, Int>) {
+private fun printScores(scores: Map<org.example.domain.Node, Double>) {
     scores.entries
         .sortedByDescending { it.value }
         .forEach { (node, score) ->
